Move pointInPolygon into a utility function.
